Added mk_object_accessor() and changed how to "reset" a translucent attribute
Since we already have class attributes and translucent attributes, we may
as well have object attributes too!  These also have default values, and
can be "reset" to them by deleting the attribute from the object.  The same
means of "resetting" makes sense for translucent attributes too (rather
than having to set them to the undefined value, as is the current means),
but requires a major $VERSION bump and appropriate notices about the change
in behaviour.